+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 
 * add isInitialized property
 * add errorMessage property
+
+## 0.0.3
+
+* change isInitialized to getter and private setter
+* change errorMessage to stored privately
+* handling access to `indicator` and `dialog` if it isn't initialized